fix: Storybookの位相付きblit契約を同期 (#62)

+++ crates/katana-ui-core-storybook/src/visual/ui_tree_canvas_scroll.rs
@@ -142,7 +142,7 @@ fn draw_visible_node(
         node,
         x,
         node_height,
-        source_y.saturating_sub(node_top),
+        source_y.saturating_sub(node_top) as f32,
         area,
         palette,
     );
